requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
tqdm>=4.64.0
matplotlib>=3.5.0
tensorboard>=2.9.0
scikit-learn>=1.0.0
PyYAML>=6.0
pandas>=1.3.0
